+++ releasenotes.md
@@ -15,6 +15,7 @@
 - Warn on if-catch with just a `default` case.
 - Compile time array inc/dec.
 - Improve error message when using ',' in struct declarations. #1920
+- Compile time array assign ops, e.g. `$c[1] += 3` #1890.
 
 ### Fixes
 - Fix issue requiring prefix on a generic interface declaration.